+++ attributes/default.rb
@@ -23,3 +23,4 @@ default['pxe_dust']['default'] = {}
 default['pxe_dust']['dir'] = '/var/www/pxe_dust'
 default['pxe_dust']['hosts_file'] = '/etc/hosts_pxe_dust'
 default['pxe_dust']['interface'] = nil
+default['pxe_dust']['port'] = 9630